am243x/am64x : pruicss_pwm : fix IepClkPeriod datatype
- convert pruIcssIepClkPeriod in PRUICSS_PWM_IEP_Attrs_s to float

Fixes: PINDSW-7513

+++ source/pruicss_pwm/include/pruicss_pwm.h
@@ -251,7 +251,7 @@ typedef struct PRUICSS_PWM_IEP_Attrs_s
     /*pruIcssIepClk frequency*/
     uint32_t pruIcssIepClkFrequency;
     /*pruIcssIepClk period in nanoseconds*/
-    uint32_t pruIcssIepClkPeriod;
+    float pruIcssIepClkPeriod;
     /*iep0 increment value*/
     uint8_t  iep0IncrementValue;
     /*enable or disable IEP0*/
